+++ dataset/README.md
@@ -1,6 +1,6 @@
 # Dataset
 
-The movieclip dataset is published to HuggingFace here - 
+The movieclip dataset is published to HuggingFace here -  
 
 This documentation shows how to generate the dataset yourself. Feel free to edit the methodology for your own purposes.
 The dataset is generated by:
@@ -14,8 +14,7 @@ The dataset is generated by:
 You need to create a .env file in this directory and populate it as follows:
 
 ```env
-TMDB_API_KEY=<your-api-key>
-TMDB_READ_TOKEN=<your-token>
+OMDB_API_KEY=<your-api-key>
 ```
 
 You can create a TMDB account [here](https://www.themoviedb.org/) and then make an API key and read token. 
@@ -39,6 +38,7 @@ To run the code for this dataset generation simply
 
 ### S3 Download
 
+_Coming soon_
 
 ## Using the Dataset
